cdo_sgups_bot/requirements.txt
aiogram==3.7.0
aiosqlite==0.20.0
groq==0.9.0
apscheduler==3.10.4
python-dotenv==1.0.1
